requirements.txt
browser-use
playwright
langchain-google-genai
langchain-ollama
asyncio
python-dotenv
streamlit